Improve logging and move shared function to helpers.go
